updated to 1.1.2

+++ libs/build_index.py
@@ -1,45 +1,44 @@
 import asyncio
 import time
 
-from datashaper import WorkflowCallbacks
 import graphrag.api as api
 import streamlit as st
 
 from libs.common import run_command, load_graphrag_config
 from libs.print_progress import PrintProgressLogger
+from theodoretools.fs import get_directory_size
 
 
 def build_index(project_name: str):
 
     if st.button("Start Build", key="build_index_" + project_name, icon="🚀"):
+        config = load_graphrag_config(project_name)
+        st.info("Config:")
+        st.json(config, expanded=False)
+
         with st.spinner("Building index..."):
-
-            progress_reporter = PrintProgressLogger("")
-            config = load_graphrag_config(project_name)
-
-            st.json(config, expanded=False)
-
-            # TODO: will support this in the future
-            # workflow_callbacks = [
-            #     WorkflowCallbacks(
-            #         on_progress=lambda progress: st.write(progress)
-            #     )
-            # ]
-
             asyncio.run(
                 api.build_index(
                     config=config,
                     run_id="",
                     is_resume_run=False,
                     memory_profile=False,
-                    progress_logger=progress_reporter,
-                    # callbacks=workflow_callbacks,
+                    progress_logger=PrintProgressLogger(""),
                 )
             )
 
-    st.markdown("----------------------------")
+    cache_size_mb = get_directory_size(f"/app/projects/{project_name}/cache")
+    if cache_size_mb > 0:
+        st.write(f"Cache size: {cache_size_mb} MB")
 
-    if st.button("Clear index files", key="clear_index_" + project_name, icon="🗑️"):
+    output_size_mb = get_directory_size(
+        f"/app/projects/{project_name}/output", [".log"]
+    )
+    if output_size_mb > 0 and st.button(
+        f"Clear index files ({output_size_mb} MB)",
+        key="clear_index_" + project_name,
+        icon="🗑️",
+    ):
         run_command(f"rm -rf /app/projects/{project_name}/output/*")
         st.success("All files deleted.")
         time.sleep(3)
